Update rmDir
Добавил селектор наличия файлов в удаляемой папке

+++ S2000/s2000.ino
@@ -145,6 +145,8 @@ String mkDir(String dir){
 String rmDir(String dir){
   File file = SD.open(myDir + "/" + dir);
   if(file){
+    File isFiles = file.openNextFile();
+    if (isFiles) return "Удаляемая папка должна быть пустой.";
     if (SD.rmdir(myDir + "/" + dir)){
       return "Папка " + dir + " удалена.";
     }
